Backend/labeling-machine/requirements.txt
requests
pynamodb
python-jose
python-ulid
requests-toolbelt
wemake-python-styleguide